python_generator/requirements.txt
openai==1.97.0
firebase-admin==6.2.0
python-dotenv==1.0.0
pydantic==2.11.7
click==8.1.7 